Add log processor
Why do we need this change?
----
Add a log processor that runs on a separate thread via watchdog.
I parses the log file and keep the stats, as well as handles alerting.
It uses a queue with 2 mins sliding windows for high traffic detection.

What effects does this change have?
----
- add log file processor
- add watcher
- add basic program scaffolding
- add basic dashboard